[REM-333] feat: add user management components and enhance user interface
- CadUser.vue: created a modal component for user registration
- TableConfig.vue: replaced data table with UsersComponent for better user management
- UsersComponent.vue: implemented user data fetching and display using DataTable
- ConfigView.vue: integrated new components into the configuration view
