+++ client/src/routes/admin/Components/CadUser.vue
@@ -1,0 +1,7 @@
+<template>
+  <BModal no-footer id="ModalFormUsr" data-bs-theme="dark" size="xl" centered class="text-white">
+    <div>
+      <BForm id="FormUsr"> </BForm>
+    </div>
+  </BModal>
+</template>
