+++ client/src/routes/admin/Components/TableConfig.vue
@@ -1,23 +1,5 @@
 <script setup>
-import { onBeforeMount, ref } from "vue";
-import { api } from "../../../main";
-import DataTable from "datatables.net-vue3";
-import DataTablesCore from "datatables.net-bs5";
-const items = ref();
-DataTable.use(DataTablesCore);
-onBeforeMount(async () => {
-  try {
-    const response = await api.get("/users");
-
-    console.log(response.data);
-
-    items.value = response.data.database.map((item) => {
-      return [item.id, item.login, item.nome_usuario, item.email];
-    });
-  } catch (error) {
-    console.error(error);
-  }
-});
+import UsersComponent from "../pages/UsersComponent.vue";
 </script>
 
 <template>
@@ -74,81 +56,7 @@ onBeforeMount(async () => {
         aria-labelledby="home-tab"
         tabindex="0"
       >
-        <div class="card" data-bs-theme="dark">
-          <div class="card-header d-grid gap-3">
-            <div class="d-flex gap-3 p-2">
-              <div class="justify-content-xxl-endalign-middle me-auto text-center">
-                <h4>Usuários</h4>
-              </div>
-              <div class="justify-content-xxl-end">
-                <a class="btn btn-outline-success me-2 fw-bold" id="download-button" href="#"
-                  >Cadastrar usuário</a
-                >
-                <button type="button" class="btn btn-outline-warning me-2 fw-bold">
-                  Encerrar Execução
-                </button>
-              </div>
-            </div>
-          </div>
-          <div class="card-body table-responsive">
-            <DataTable :data="items" class="placeholder-glow table table-striped table-hover">
-              <thead>
-                <tr>
-                  <th>#</th>
-                  <th>Login</th>
-                  <th>Nome</th>
-                  <th>E-mail</th>
-                </tr>
-              </thead>
-              <!-- <template #column-7="props">
-                <button
-                  v-if="props.rowData[5].toString().toLowerCase() !== 'em execução'"
-                  class="btn btn-sm btn-success"
-                  data-bs-toggle="tooltip"
-                  data-bs-title="Default tooltip"
-                  @click="download_file(props.cellData)"
-                >
-                  <FontAwesomeIcon :icon="faDownload" />
-                </button>
-
-                <a
-                  v-else-if="props.rowData[5].toString().toLowerCase() === 'em execução'"
-                  :href="`/logs_bot/${props.rowData[0]}`"
-                  class="btn btn-sm btn-primary"
-                  data-bs-toggle="tooltip"
-                  data-bs-title="Default tooltip"
-                  ><FontAwesomeIcon :icon="faEye"
-                /></a>
-              </template> -->
-              <tfoot>
-                <tr>
-                  <th>#</th>
-                  <th>Login</th>
-                  <th>Nome</th>
-                  <th>E-mail</th>
-                </tr>
-              </tfoot>
-            </DataTable>
-            <!-- <table class="table table-striped table-hover">
-              <thead>
-                <tr>
-                  <th scope="col">#</th>
-                  <th scope="col">Login</th>
-                  <th scope="col">Nome</th>
-                  <th scope="col">E-Mail</th>
-                </tr>
-              </thead>
-              <tbody v-for="item in items" :key="item.id">
-                <tr>
-                  <th scope="row">{{ item.id }}</th>
-                  <td>{{ item.login }}</td>
-                  <td>{{ item.nome_usuario }}</td>
-                  <td>{{ item.email }}</td>
-                </tr>
-              </tbody>
-            </table> -->
-          </div>
-        </div>
+        <UsersComponent />
       </div>
       <div
         class="tab-pane fade"
